pyproject.toml
[tool.poetry]
name = "nba-live"
version = "0.1.0"
description = "Local NBA database ETL & AI agent"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2"
sqlalchemy = "^2.0"
duckdb = "^0.10"
langchain = "^0.1.16"
openai = "^1.25"
pydantic = "^2.7"
python-dotenv = "^1.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1"
ruff = "^0.4"
black = "^24.4"
mypy = "^1.9"
pre-commit = "^3.7"

[tool.ruff]
line-length = 88

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api" 